# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Neon Murer CMS API
  description: |
    Vollständige REST API für das Neon Murer Content Management System.
    
    ## Authentifizierung
    Die meisten Endpoints erfordern JWT-Token Authentication über den `Authorization: Bearer <token>` Header.
    
    ## Fehlerbehandlung
    Alle Endpoints verwenden einheitliche HTTP-Status-Codes und JSON-Antworten.
    
    ## Rate Limiting
    API-Aufrufe sind auf 100 Requests pro Minute pro IP-Adresse begrenzt.
    
  version: 1.0.0
  contact:
    name: Neon Murer AG
    url: https://neonmurer.ch
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api
    description: Development Server
  - url: https://cms.neonmurer.ch/api  
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Fehlermeldung"
        details:
          type: object
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validierungsfehler"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    User:
      type: object
      properties:
        id:
          type: string
          example: "cmdid4eas0000wa04ime8men5"
        email:
          type: string
          format: email
          example: "admin@neonmurer.ch"
        firstName:
          type: string
          example: "Max"
        lastName:
          type: string
          example: "Mustermann"
        role:
          type: string
          enum: [admin, editor, viewer]
          example: "admin"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          example: "def456789abcdef..."
        expiresIn:
          type: integer
          example: 3600
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Benutzer anmelden
      description: Meldet einen Benutzer an und gibt JWT-Tokens zurück
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "admin@neonmurer.ch"
                password:
                  type: string
                  example: "securePassword123"
      responses:
        '200':
          description: Erfolgreich angemeldet
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Login erfolgreich"
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Ungültige Anmeldedaten
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validierungsfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /auth/register:
    post:
      tags: [Auth]
      summary: Neuen Benutzer registrieren
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                firstName:
                  type: string
                lastName:
                  type: string
                role:
                  type: string
                  enum: [admin, editor, viewer]
                  default: viewer
      responses:
        '201':
          description: Benutzer erfolgreich erstellt
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Benutzer erfolgreich erstellt"
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '409':
          description: E-Mail bereits vergeben
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags: [Auth]
      summary: Aktuelle Benutzerinfo abrufen
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Benutzerinfo erfolgreich abgerufen
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Nicht authentifiziert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Access Token erneuern
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token erfolgreich erneuert
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Ungültiger Refresh Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Benutzer abmelden
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Erfolgreich abgemeldet
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Logout erfolgreich"
                    
  # ANALYTICS ENDPOINTS
...